[NUI.Components]Improve SAM score of NUI.Components
