+++ src/Tizen.NUI.Components/Controls/DropDown.DropDownItemView.cs
@@ -1,0 +1,306 @@
+﻿using System.ComponentModel;
+using Tizen.NUI.BaseComponents;
+
+namespace Tizen.NUI.Components
+{
+    public partial class DropDown
+    {
+        /// This will be public opened in tizen_5.5 after ACR done. Before ACR, need to be hidden as inhouse API.
+        [EditorBrowsable(EditorBrowsableState.Never)]
+        internal class DropDownItemView : Control
+        {
+            private TextLabel mText = null;
+            private ImageView mIcon = null;
+            private ImageView mCheck = null;
+
+            /// This will be public opened in tizen_5.5 after ACR done. Before ACR, need to be hidden as inhouse API.
+            [EditorBrowsable(EditorBrowsableState.Never)]
+            public DropDownItemView() : base() { }
+
+            /// This will be public opened in tizen_5.5 after ACR done. Before ACR, need to be hidden as inhouse API.
+            [EditorBrowsable(EditorBrowsableState.Never)]
+            public Selector<Color> BackgroundColorSelector { get; set; }
+
+            /// This will be public opened in tizen_5.5 after ACR done. Before ACR, need to be hidden as inhouse API.
+            [EditorBrowsable(EditorBrowsableState.Never)]
+            public string Text
+            {
+                get
+                {
+                    return (null == mText) ? null : mText.Text;
+                }
+                set
+                {
+                    CreateText();
+                    mText.Text = value;
+                }
+            }
+
+            /// This will be public opened in tizen_5.5 after ACR done. Before ACR, need to be hidden as inhouse API.
+            [EditorBrowsable(EditorBrowsableState.Never)]
+            public string FontFamily
+            {
+                get
+                {
+                    return (null == mText) ? null : mText.FontFamily;
+                }
+                set
+                {
+                    CreateText();
+                    mText.FontFamily = value;
+                }
+            }
+
+            /// This will be public opened in tizen_5.5 after ACR done. Before ACR, need to be hidden as inhouse API.
+            [EditorBrowsable(EditorBrowsableState.Never)]
+            public float? PointSize
+            {
+                get
+                {
+                    return (null == mText) ? 0 : mText.PointSize;
+                }
+                set
+                {
+                    CreateText();
+                    mText.PointSize = (float)value;
+                }
+            }
+
+            /// This will be public opened in tizen_5.5 after ACR done. Before ACR, need to be hidden as inhouse API.
+            [EditorBrowsable(EditorBrowsableState.Never)]
+            public Color TextColor
+            {
+                get
+                {
+                    return (null == mText) ? null : mText.TextColor;
+                }
+                set
+                {
+                    CreateText();
+                    mText.TextColor = value;
+                }
+            }
+
+            /// This will be public opened in tizen_5.5 after ACR done. Before ACR, need to be hidden as inhouse API.
+            [EditorBrowsable(EditorBrowsableState.Never)]
+            public Position TextPosition
+            {
+                get
+                {
+                    return (null == mText) ? null : mText.Position;
+                }
+                set
+                {
+                    CreateText();
+                    mText.Position = value;
+                }
+            }
+
+            /// This will be public opened in tizen_5.5 after ACR done. Before ACR, need to be hidden as inhouse API.
+            [EditorBrowsable(EditorBrowsableState.Never)]
+            public string IconResourceUrl
+            {
+                get
+                {
+                    return (null == mIcon) ? null : mIcon.ResourceUrl;
+                }
+                set
+                {
+                    CreateIcon();
+                    mIcon.ResourceUrl = value;
+                }
+            }
+
+            /// This will be public opened in tizen_5.5 after ACR done. Before ACR, need to be hidden as inhouse API.
+            [EditorBrowsable(EditorBrowsableState.Never)]
+            public Size IconSize
+            {
+                get
+                {
+                    return (null == mIcon) ? null : mIcon.Size;
+                }
+                set
+                {
+                    CreateIcon();
+                    mIcon.Size = value;
+                }
+            }
+
+            /// This will be public opened in tizen_5.5 after ACR done. Before ACR, need to be hidden as inhouse API.
+            [EditorBrowsable(EditorBrowsableState.Never)]
+            public Position IconPosition
+            {
+                get
+                {
+                    return (null == mIcon) ? null : mIcon.Position;
+                }
+                set
+                {
+                    CreateIcon();
+                    mIcon.Position = value;
+                }
+            }
+
+            /// This will be public opened in tizen_5.5 after ACR done. Before ACR, need to be hidden as inhouse API.
+            [EditorBrowsable(EditorBrowsableState.Never)]
+            public string CheckResourceUrl
+            {
+                get
+                {
+                    return (null == mCheck) ? null : mCheck.ResourceUrl;
+                }
+                set
+                {
+                    CreateCheckImage();
+                    mCheck.ResourceUrl = value;
+                }
+            }
+
+            /// This will be public opened in tizen_5.5 after ACR done. Before ACR, need to be hidden as inhouse API.
+            [EditorBrowsable(EditorBrowsableState.Never)]
+            public Position CheckPosition
+            {
+                get
+                {
+                    return (null == mCheck) ? null : mCheck.Position;
+                }
+                set
+                {
+                    CreateCheckImage();
+                    mCheck.Position = value;
+                }
+            }
+
+            /// This will be public opened in tizen_5.5 after ACR done. Before ACR, need to be hidden as inhouse API.
+            [EditorBrowsable(EditorBrowsableState.Never)]
+            public Size CheckImageSize
+            {
+                get
+                {
+                    return (null == mCheck) ? null : mCheck.Size;
+                }
+                set
+                {
+                    CreateCheckImage();
+                    mCheck.Size = value;
+                }
+            }
+
+            /// This will be public opened in tizen_5.5 after ACR done. Before ACR, need to be hidden as inhouse API.
+            [EditorBrowsable(EditorBrowsableState.Never)]
+            public bool IsSelected
+            {
+                get
+                {
+                    return (null == mCheck) ? false : mCheck.Visibility;
+                }
+                set
+                {
+                    CreateCheckImage();
+                    if (value)
+                    {
+                        mCheck.Show();
+                    }
+                    else
+                    {
+                        mCheck.Hide();
+                    }
+                }
+            }
+
+            /// This will be public opened in tizen_5.5 after ACR done. Before ACR, need to be hidden as inhouse API.
+            [EditorBrowsable(EditorBrowsableState.Never)]
+            protected override void Dispose(DisposeTypes type)
+            {
+                if (disposed)
+                {
+                    return;
+                }
+
+                if (type == DisposeTypes.Explicit)
+                {
+                    if (mText != null)
+                    {
+                        Remove(mText);
+                        mText.Dispose();
+                        mText = null;
+                    }
+
+                    if (mIcon != null)
+                    {
+                        Remove(mIcon);
+                        mIcon.Dispose();
+                        mIcon = null;
+                    }
+
+                    if (mCheck != null)
+                    {
+                        Remove(mCheck);
+                        mCheck.Dispose();
+                        mCheck = null;
+                    }
+                }
+                base.Dispose(type);
+            }
+
+            /// <summary>
+            /// Get DropDownItemView style.
+            /// </summary>
+            /// <returns>The empty.</returns>
+            /// This will be public opened in tizen_5.5 after ACR done. Before ACR, need to be hidden as inhouse API.
+            [EditorBrowsable(EditorBrowsableState.Never)]
+            protected override ViewStyle CreateViewStyle()
+            {
+                return new DropDownItemStyle();
+            }
+
+            private void CreateIcon()
+            {
+                if (mIcon == null)
+                {
+                    mIcon = new ImageView()
+                    {
+                        PositionUsesPivotPoint = true,
+                        ParentOrigin = Tizen.NUI.ParentOrigin.TopLeft,
+                        PivotPoint = Tizen.NUI.PivotPoint.TopLeft,
+                    };
+                    Add(mIcon);
+                }
+            }
+
+            private void CreateText()
+            {
+                if (mText == null)
+                {
+                    mText = new TextLabel()
+                    {
+                        PositionUsesPivotPoint = true,
+                        ParentOrigin = Tizen.NUI.ParentOrigin.TopLeft,
+                        PivotPoint = Tizen.NUI.PivotPoint.TopLeft,
+                        WidthResizePolicy = ResizePolicyType.UseNaturalSize,
+                        HeightResizePolicy = ResizePolicyType.FillToParent,
+                        VerticalAlignment = VerticalAlignment.Center,
+                        HorizontalAlignment = HorizontalAlignment.Begin,
+                    };
+                    Add(mText);
+                }
+            }
+
+            private void CreateCheckImage()
+            {
+                if (mCheck == null)
+                {
+                    mCheck = new ImageView()
+                    {
+                        PositionUsesPivotPoint = true,
+                        ParentOrigin = Tizen.NUI.ParentOrigin.TopLeft,
+                        PivotPoint = Tizen.NUI.PivotPoint.TopLeft,
+                        Name = "checkedImage",
+                    };
+                    Add(mCheck);
+                }
+                mCheck.Hide();
+            }
+        }
+    }
+}
