+++ src/Tizen.NUI.Components/Controls/Slider.Internal.cs
@@ -1,0 +1,881 @@
+﻿using System;
+using System.ComponentModel;
+using Tizen.NUI.BaseComponents;
+
+namespace Tizen.NUI.Components
+{
+    public partial class Slider
+    {
+
+        /// <summary>
+        /// Get Slider style.
+        /// </summary>
+        /// <returns>The default slider style.</returns>
+        /// <since_tizen> 8 </since_tizen>
+        protected override ViewStyle CreateViewStyle()
+        {
+            return new SliderStyle();
+        }
+
+        /// <summary>
+        /// Dispose Slider.
+        /// </summary>
+        /// <param name="type">Dispose type.</param>
+        /// <since_tizen> 6 </since_tizen>
+        protected override void Dispose(DisposeTypes type)
+        {
+            if (disposed)
+            {
+                return;
+            }
+
+            if (type == DisposeTypes.Explicit)
+            {
+                if (null != panGestureDetector)
+                {
+                    if (null != thumbImage)
+                    {
+                        panGestureDetector.Detach(thumbImage);
+                    }
+                    panGestureDetector.Detected -= OnPanGestureDetected;
+                    panGestureDetector.Dispose();
+                    panGestureDetector = null;
+                }
+
+                if (null != thumbImage)
+                {
+                    thumbImage.TouchEvent -= OnTouchEventForThumb;
+                    Utility.Dispose(thumbImage);
+                }
+                Utility.Dispose(slidedTrackImage);
+                if (null != bgTrackImage)
+                {
+                    bgTrackImage.TouchEvent -= OnTouchEventForBgTrack;
+                    Utility.Dispose(bgTrackImage);
+                }
+                Utility.Dispose(lowIndicatorImage);
+                Utility.Dispose(highIndicatorImage);
+                Utility.Dispose(lowIndicatorText);
+                Utility.Dispose(highIndicatorText);
+            }
+
+            base.Dispose(type);
+        }
+
+        /// <summary>
+        /// Update Slider by style.
+        /// </summary>
+        /// <since_tizen> 6 </since_tizen>
+        /// This will be public opened in tizen_5.5 after ACR done. Before ACR, need to be hidden as inhouse API.
+        [EditorBrowsable(EditorBrowsableState.Never)]
+        protected override void OnUpdate()
+        {
+            RelayoutBaseComponent();
+
+            UpdateComponentByIndicatorTypeChanged();
+            UpdateBgTrackSize();
+            UpdateBgTrackPosition();
+            UpdateLowIndicatorSize();
+            UpdateValue();
+        }
+
+        /// <summary>
+        /// Theme change callback when theme is changed, this callback will be trigger.
+        /// </summary>
+        /// <param name="sender">The sender</param>
+        /// <param name="e">The event data</param>
+        /// <since_tizen> 8 </since_tizen>
+        protected override void OnThemeChangedEvent(object sender, StyleManager.ThemeChangeEventArgs e)
+        {
+            SliderStyle sliderStyle = StyleManager.Instance.GetViewStyle(StyleName) as SliderStyle;
+            if (sliderStyle != null)
+            {
+                Style?.CopyFrom(sliderStyle);
+                RelayoutRequest();
+            }
+        }
+
+
+        private void Initialize()
+        {
+            currentSlidedOffset = 0;
+            isFocused = false;
+            isPressed = false;
+            LayoutDirectionChanged += OnLayoutDirectionChanged;
+        }
+
+        private void OnLayoutDirectionChanged(object sender, LayoutDirectionChangedEventArgs e)
+        {
+            RelayoutRequest();
+        }
+
+        private ImageView CreateSlidedTrack()
+        {
+            if (null == slidedTrackImage)
+            {
+                slidedTrackImage = new ImageView()
+                {
+                    WidthResizePolicy = ResizePolicyType.Fixed,
+                    HeightResizePolicy = ResizePolicyType.Fixed
+                };
+
+                if (bgTrackImage != null)
+                {
+                    bgTrackImage.Add(slidedTrackImage);
+                }
+
+                if (null != thumbImage)
+                {
+                    slidedTrackImage.Add(thumbImage);
+                }
+            }
+
+            return slidedTrackImage;
+        }
+
+        private TextLabel CreateLowIndicatorText()
+        {
+            if (null == lowIndicatorText)
+            {
+                lowIndicatorText = new TextLabel()
+                {
+                    WidthResizePolicy = ResizePolicyType.Fixed,
+                    HeightResizePolicy = ResizePolicyType.Fixed
+                };
+                this.Add(lowIndicatorText);
+            }
+
+            return lowIndicatorText;
+        }
+
+        private TextLabel CreateHighIndicatorText()
+        {
+            if (null == highIndicatorText)
+            {
+                highIndicatorText = new TextLabel()
+                {
+                    WidthResizePolicy = ResizePolicyType.Fixed,
+                    HeightResizePolicy = ResizePolicyType.Fixed
+                };
+                this.Add(highIndicatorText);
+            }
+
+            return highIndicatorText;
+        }
+
+        private ImageView CreateBackgroundTrack()
+        {
+            if (null == bgTrackImage)
+            {
+                bgTrackImage = new ImageView()
+                {
+                    WidthResizePolicy = ResizePolicyType.Fixed,
+                    HeightResizePolicy = ResizePolicyType.Fixed,
+                    ParentOrigin = Tizen.NUI.ParentOrigin.Center,
+                    PivotPoint = Tizen.NUI.PivotPoint.Center,
+                    PositionUsesPivotPoint = true
+                };
+                this.Add(bgTrackImage);
+
+                if (null != slidedTrackImage)
+                {
+                    bgTrackImage.Add(slidedTrackImage);
+                }
+
+                bgTrackImage.TouchEvent += OnTouchEventForBgTrack;
+            }
+
+            return bgTrackImage;
+        }
+
+        private ImageView CreateThumb()
+        {
+            if (null == thumbImage)
+            {
+                thumbImage = new ImageView()
+                {
+                    WidthResizePolicy = ResizePolicyType.Fixed,
+                    HeightResizePolicy = ResizePolicyType.Fixed,
+                    ParentOrigin = NUI.ParentOrigin.Center,
+                    PivotPoint = NUI.PivotPoint.Center,
+                    PositionUsesPivotPoint = true
+                };
+                if (slidedTrackImage != null)
+                {
+                    slidedTrackImage.Add(thumbImage);
+                }
+                thumbImage.TouchEvent += OnTouchEventForThumb;
+
+                panGestureDetector = new PanGestureDetector();
+                panGestureDetector.Attach(thumbImage);
+                panGestureDetector.Detected += OnPanGestureDetected;
+            }
+
+            return thumbImage;
+        }
+
+        private void OnPanGestureDetected(object source, PanGestureDetector.DetectedEventArgs e)
+        {
+            if (e.PanGesture.State == Gesture.StateType.Started)
+            {
+                if (direction == DirectionType.Horizontal)
+                {
+                    currentSlidedOffset = slidedTrackImage.SizeWidth;
+                }
+                else if (direction == DirectionType.Vertical)
+                {
+                    currentSlidedOffset = slidedTrackImage.SizeHeight;
+                }
+                UpdateState(isFocused, true);
+            }
+
+            if (e.PanGesture.State == Gesture.StateType.Continuing || e.PanGesture.State == Gesture.StateType.Started)
+            {
+                if (direction == DirectionType.Horizontal)
+                {
+                    CalculateCurrentValueByGesture(e.PanGesture.Displacement.X);
+                }
+                else if (direction == DirectionType.Vertical)
+                {
+                    CalculateCurrentValueByGesture(-e.PanGesture.Displacement.Y);
+                }
+                UpdateValue();
+            }
+
+            if (e.PanGesture.State == Gesture.StateType.Finished)
+            {
+                if (null != slidingFinishedHandler)
+                {
+                    SlidingFinishedArgs args = new SlidingFinishedArgs();
+                    args.CurrentValue = curValue;
+                    slidingFinishedHandler(this, args);
+                }
+
+                UpdateState(isFocused, false);
+            }
+        }
+
+        // Relayout basic component: track, thumb and indicator
+        private void RelayoutBaseComponent(bool isInitial = true)
+        {
+            if (direction == DirectionType.Horizontal)
+            {
+                if (slidedTrackImage != null)
+                {
+                    slidedTrackImage.ParentOrigin = NUI.ParentOrigin.CenterLeft;
+                    slidedTrackImage.PivotPoint = NUI.PivotPoint.CenterLeft;
+                    slidedTrackImage.PositionUsesPivotPoint = true;
+                }
+                if (thumbImage != null)
+                {
+                    thumbImage.ParentOrigin = NUI.ParentOrigin.CenterRight;
+                    thumbImage.PivotPoint = NUI.PivotPoint.Center;
+                    thumbImage.PositionUsesPivotPoint = true;
+                }
+                if (lowIndicatorImage != null)
+                {
+                    lowIndicatorImage.ParentOrigin = NUI.ParentOrigin.CenterLeft;
+                    lowIndicatorImage.PivotPoint = NUI.PivotPoint.CenterLeft;
+                    lowIndicatorImage.PositionUsesPivotPoint = true;
+                }
+                if (highIndicatorImage != null)
+                {
+                    highIndicatorImage.ParentOrigin = NUI.ParentOrigin.CenterRight;
+                    highIndicatorImage.PivotPoint = NUI.PivotPoint.CenterRight;
+                    highIndicatorImage.PositionUsesPivotPoint = true;
+                }
+                if (lowIndicatorText != null)
+                {
+                    lowIndicatorText.ParentOrigin = NUI.ParentOrigin.CenterLeft;
+                    lowIndicatorText.PivotPoint = NUI.PivotPoint.CenterLeft;
+                    lowIndicatorText.PositionUsesPivotPoint = true;
+                }
+                if (highIndicatorText != null)
+                {
+                    highIndicatorText.ParentOrigin = NUI.ParentOrigin.CenterRight;
+                    highIndicatorText.PivotPoint = NUI.PivotPoint.CenterRight;
+                    highIndicatorText.PositionUsesPivotPoint = true;
+                }
+                if (panGestureDetector != null)
+                {
+                    if (!isInitial)
+                    {
+                        panGestureDetector.RemoveDirection(PanGestureDetector.DirectionVertical);
+                    }
+                    panGestureDetector.AddDirection(PanGestureDetector.DirectionHorizontal);
+                }
+            }
+            else if (direction == DirectionType.Vertical)
+            {
+                if (slidedTrackImage != null)
+                {
+                    slidedTrackImage.ParentOrigin = NUI.ParentOrigin.BottomCenter;
+                    slidedTrackImage.PivotPoint = NUI.PivotPoint.BottomCenter;
+                    slidedTrackImage.PositionUsesPivotPoint = true;
+                }
+                if (thumbImage != null)
+                {
+                    thumbImage.ParentOrigin = NUI.ParentOrigin.TopCenter;
+                    thumbImage.PivotPoint = NUI.PivotPoint.Center;
+                    thumbImage.PositionUsesPivotPoint = true;
+                }
+                if (lowIndicatorImage != null)
+                {
+                    lowIndicatorImage.ParentOrigin = NUI.ParentOrigin.BottomCenter;
+                    lowIndicatorImage.PivotPoint = NUI.PivotPoint.BottomCenter;
+                    lowIndicatorImage.PositionUsesPivotPoint = true;
+                }
+                if (highIndicatorImage != null)
+                {
+                    highIndicatorImage.ParentOrigin = NUI.ParentOrigin.TopCenter;
+                    highIndicatorImage.PivotPoint = NUI.PivotPoint.TopCenter;
+                    highIndicatorImage.PositionUsesPivotPoint = true;
+                }
+                if (lowIndicatorText != null)
+                {
+                    lowIndicatorText.ParentOrigin = NUI.ParentOrigin.BottomCenter;
+                    lowIndicatorText.PivotPoint = NUI.PivotPoint.BottomCenter;
+                    lowIndicatorText.PositionUsesPivotPoint = true;
+                }
+                if (highIndicatorText != null)
+                {
+                    highIndicatorText.ParentOrigin = NUI.ParentOrigin.TopCenter;
+                    highIndicatorText.PivotPoint = NUI.PivotPoint.TopCenter;
+                    highIndicatorText.PositionUsesPivotPoint = true;
+                }
+                if (panGestureDetector != null)
+                {
+                    if (!isInitial)
+                    {
+                        panGestureDetector.RemoveDirection(PanGestureDetector.DirectionHorizontal);
+                    }
+                    panGestureDetector.AddDirection(PanGestureDetector.DirectionVertical);
+                }
+            }
+        }
+
+        private int BgTrackLength()
+        {
+            int bgTrackLength = 0;
+            IndicatorType type = CurrentIndicatorType();
+
+            if (type == IndicatorType.None)
+            {
+                if (direction == DirectionType.Horizontal)
+                {
+                    bgTrackLength = this.Size2D.Width;
+                }
+                else if (direction == DirectionType.Vertical)
+                {
+                    bgTrackLength = this.Size2D.Height;
+                }
+            }
+            else if (type == IndicatorType.Image)
+            {// <lowIndicatorImage> <spaceBetweenTrackAndIndicator> <bgTrack> <spaceBetweenTrackAndIndicator> <highIndicatorImage>
+                Size lowIndicatorImageSize = LowIndicatorImageSize();
+                Size highIndicatorImageSize = HighIndicatorImageSize();
+                int curSpace = (int)CurrentSpaceBetweenTrackAndIndicator();
+                if (direction == DirectionType.Horizontal)
+                {
+                    int lowIndicatorSpace = ((lowIndicatorImageSize.Width == 0) ? (0) : ((int)(curSpace + lowIndicatorImageSize.Width)));
+                    int highIndicatorSpace = ((highIndicatorImageSize.Width == 0) ? (0) : ((int)(curSpace + highIndicatorImageSize.Width)));
+                    bgTrackLength = this.Size2D.Width - lowIndicatorSpace - highIndicatorSpace;
+                }
+                else if (direction == DirectionType.Vertical)
+                {
+                    int lowIndicatorSpace = ((lowIndicatorImageSize.Height == 0) ? (0) : ((int)(curSpace + lowIndicatorImageSize.Height)));
+                    int highIndicatorSpace = ((highIndicatorImageSize.Height == 0) ? (0) : ((int)(curSpace + highIndicatorImageSize.Height)));
+                    bgTrackLength = this.Size2D.Height - lowIndicatorSpace - highIndicatorSpace;
+                }
+            }
+            else if (type == IndicatorType.Text)
+            {// <lowIndicatorText> <spaceBetweenTrackAndIndicator> <bgTrack> <spaceBetweenTrackAndIndicator> <highIndicatorText>
+                Size lowIndicatorTextSize = LowIndicatorTextSize();
+                Size highIndicatorTextSize = HighIndicatorTextSize();
+                int curSpace = (int)CurrentSpaceBetweenTrackAndIndicator();
+                if (direction == DirectionType.Horizontal)
+                {
+                    int lowIndicatorSpace = ((lowIndicatorTextSize.Width == 0) ? (0) : ((int)(curSpace + lowIndicatorTextSize.Width)));
+                    int highIndicatorSpace = ((highIndicatorTextSize.Width == 0) ? (0) : ((int)(curSpace + highIndicatorTextSize.Width)));
+                    bgTrackLength = this.Size2D.Width - lowIndicatorSpace - highIndicatorSpace;
+                }
+                else if (direction == DirectionType.Vertical)
+                {
+                    int lowIndicatorSpace = ((lowIndicatorTextSize.Height == 0) ? (0) : ((int)(curSpace + lowIndicatorTextSize.Height)));
+                    int highIndicatorSpace = ((highIndicatorTextSize.Height == 0) ? (0) : ((int)(curSpace + highIndicatorTextSize.Height)));
+                    bgTrackLength = this.Size2D.Height - lowIndicatorSpace - highIndicatorSpace;
+                }
+            }
+            return bgTrackLength;
+        }
+
+        private void UpdateLowIndicatorSize()
+        {
+            if (lowIndicatorSize != null)
+            {
+                if (lowIndicatorImage != null)
+                {
+                    lowIndicatorImage.Size = lowIndicatorSize;
+                }
+                if (lowIndicatorText != null)
+                {
+                    lowIndicatorText.Size = lowIndicatorSize;
+                }
+            }
+            else
+            {
+                if (lowIndicatorImage != null && Style != null && Style.LowIndicatorImage != null && Style.LowIndicatorImage.Size != null)
+                {
+                    lowIndicatorImage.Size = Style.LowIndicatorImage.Size;
+                }
+                if (lowIndicatorText != null && Style != null && Style.LowIndicator != null && Style.LowIndicator.Size != null)
+                {
+                    lowIndicatorText.Size = Style.LowIndicator.Size;
+                }
+            }
+        }
+
+        private void UpdateBgTrackSize()
+        {
+            if (bgTrackImage == null)
+            {
+                return;
+            }
+            int curTrackThickness = (int)CurrentTrackThickness();
+            int bgTrackLength = BgTrackLength();
+            if (direction == DirectionType.Horizontal)
+            {
+                bgTrackImage.Size2D = new Size2D(bgTrackLength, curTrackThickness);
+            }
+            else if (direction == DirectionType.Vertical)
+            {
+                bgTrackImage.Size2D = new Size2D(curTrackThickness, bgTrackLength);
+            }
+        }
+
+        private void UpdateBgTrackPosition()
+        {
+            if (bgTrackImage == null)
+            {
+                return;
+            }
+            IndicatorType type = CurrentIndicatorType();
+
+            if (type == IndicatorType.None)
+            {
+                bgTrackImage.Position2D = new Position2D(0, 0);
+            }
+            else if (type == IndicatorType.Image)
+            {
+                Size lowIndicatorImageSize = LowIndicatorImageSize();
+                Size highIndicatorImageSize = HighIndicatorImageSize();
+                int curSpace = (int)CurrentSpaceBetweenTrackAndIndicator();
+                if (direction == DirectionType.Horizontal)
+                {
+                    int lowIndicatorSpace = ((lowIndicatorImageSize.Width == 0) ? (0) : ((int)(curSpace + lowIndicatorImageSize.Width)));
+                    int highIndicatorSpace = ((highIndicatorImageSize.Width == 0) ? (0) : ((int)(curSpace + highIndicatorImageSize.Width)));
+                    bgTrackImage.Position2D = new Position2D(lowIndicatorSpace - (lowIndicatorSpace + highIndicatorSpace) / 2, 0);
+                }
+                else if (direction == DirectionType.Vertical)
+                {
+                    int lowIndicatorSpace = ((lowIndicatorImageSize.Height == 0) ? (0) : ((int)(curSpace + lowIndicatorImageSize.Height)));
+                    int highIndicatorSpace = ((highIndicatorImageSize.Height == 0) ? (0) : ((int)(curSpace + highIndicatorImageSize.Height)));
+                    bgTrackImage.Position2D = new Position2D(0, lowIndicatorSpace - (lowIndicatorSpace + highIndicatorSpace) / 2);
+                }
+            }
+            else if (type == IndicatorType.Text)
+            {
+                Size lowIndicatorTextSize = LowIndicatorTextSize();
+                Size highIndicatorTextSize = HighIndicatorTextSize();
+                int curSpace = (int)CurrentSpaceBetweenTrackAndIndicator();
+                if (direction == DirectionType.Horizontal)
+                {
+                    int lowIndicatorSpace = ((lowIndicatorTextSize.Width == 0) ? (0) : ((int)(curSpace + lowIndicatorTextSize.Width)));
+                    int highIndicatorSpace = ((highIndicatorTextSize.Width == 0) ? (0) : ((int)(curSpace + highIndicatorTextSize.Width)));
+                    bgTrackImage.Position2D = new Position2D(lowIndicatorSpace - (lowIndicatorSpace + highIndicatorSpace) / 2, 0);
+                }
+                else if (direction == DirectionType.Vertical)
+                {
+                    int lowIndicatorSpace = ((lowIndicatorTextSize.Height == 0) ? (0) : ((int)(curSpace + lowIndicatorTextSize.Height)));
+                    int highIndicatorSpace = ((highIndicatorTextSize.Height == 0) ? (0) : ((int)(curSpace + highIndicatorTextSize.Height)));
+                    bgTrackImage.Position2D = new Position2D(0, -(lowIndicatorSpace - (lowIndicatorSpace + highIndicatorSpace) / 2));
+                }
+            }
+        }
+
+        private void UpdateValue()
+        {
+            if (slidedTrackImage == null)
+            {
+                return;
+            }
+            if (curValue < minValue || curValue > maxValue || minValue >= maxValue)
+            {
+                return;
+            }
+
+            float ratio = 0;
+            ratio = (float)(curValue - minValue) / (float)(maxValue - minValue);
+
+            uint curTrackThickness = CurrentTrackThickness();
+
+            if (direction == DirectionType.Horizontal)
+            {
+                if (LayoutDirection == ViewLayoutDirectionType.RTL)
+                {
+                    ratio = 1.0f - ratio;
+                }
+                float slidedTrackLength = (float)BgTrackLength() * ratio;
+                slidedTrackImage.Size2D = new Size2D((int)(slidedTrackLength + round), (int)curTrackThickness); //Add const round to reach Math.Round function.
+            }
+            else if (direction == DirectionType.Vertical)
+            {
+                float slidedTrackLength = (float)BgTrackLength() * ratio;
+                slidedTrackImage.Size2D = new Size2D((int)(curTrackThickness + round), (int)slidedTrackLength); //Add const round to reach Math.Round function.
+            }
+        }
+
+        private uint CurrentTrackThickness()
+        {
+            uint curTrackThickness = 0;
+            if (trackThickness != null)
+            {
+                curTrackThickness = trackThickness.Value;
+            }
+            else
+            {
+                if (Style != null && Style.TrackThickness != null)
+                {
+                    curTrackThickness = Style.TrackThickness.Value;
+                }
+            }
+            return curTrackThickness;
+        }
+
+        private uint CurrentSpaceBetweenTrackAndIndicator()
+        {
+            uint curSpace = 0;
+            if (spaceBetweenTrackAndIndicator != null)
+            {
+                curSpace = spaceBetweenTrackAndIndicator.Start;
+            }
+            else
+            {
+                if (Style != null && Style.TrackPadding != null)
+                {
+                    curSpace = Style.TrackPadding.Start;
+                }
+            }
+            return curSpace;
+        }
+
+        private IndicatorType CurrentIndicatorType()
+        {
+            IndicatorType type = IndicatorType.None;
+            if (Style != null)
+            {
+                type = (IndicatorType)Style.IndicatorType;
+            }
+            return type;
+        }
+
+        private Size LowIndicatorImageSize()
+        {
+            Size size = new Size(0, 0);
+            if (lowIndicatorSize != null)
+            {
+                size = lowIndicatorSize;
+            }
+            else
+            {
+                if (Style != null && Style.LowIndicatorImage != null && Style.LowIndicatorImage.Size != null)
+                {
+                    size = Style.LowIndicatorImage.Size;
+                }
+            }
+            return size;
+        }
+
+        private Size HighIndicatorImageSize()
+        {
+            Size size = new Size(0, 0);
+            if (highIndicatorSize != null)
+            {
+                size = highIndicatorSize;
+            }
+            else
+            {
+                if (Style != null && Style.HighIndicatorImage != null && Style.HighIndicatorImage.Size != null)
+                {
+                    size = Style.HighIndicatorImage.Size;
+                }
+            }
+            return size;
+        }
+
+        private Size LowIndicatorTextSize()
+        {
+            Size size = new Size(0, 0);
+            if (lowIndicatorSize != null)
+            {
+                size = lowIndicatorSize;
+            }
+            else
+            {
+                if (Style != null && Style.LowIndicator != null && Style.LowIndicator.Size != null)
+                {
+                    size = Style.LowIndicator.Size;
+                }
+            }
+            return size;
+        }
+
+        private Size HighIndicatorTextSize()
+        {
+            Size size = new Size(0, 0);
+            if (highIndicatorSize != null)
+            {
+                size = highIndicatorSize;
+            }
+            else
+            {
+                if (Style != null && Style.HighIndicator != null && Style.HighIndicator.Size != null)
+                {
+                    size = Style.HighIndicator.Size;
+                }
+            }
+            return size;
+        }
+
+        private void CalculateCurrentValueByGesture(float offset)
+        {
+            currentSlidedOffset += offset;
+
+            if (currentSlidedOffset <= 0)
+            {
+                curValue = minValue;
+            }
+            else if (currentSlidedOffset >= BgTrackLength())
+            {
+                curValue = maxValue;
+            }
+            else
+            {
+                int bgTrackLength = BgTrackLength();
+                if (bgTrackLength != 0)
+                {
+                    curValue = ((currentSlidedOffset / (float)bgTrackLength) * (float)(maxValue - minValue)) + minValue;
+                }
+            }
+            if (valueChangedHandler != null)
+            {
+                ValueChangedArgs args = new ValueChangedArgs();
+                args.CurrentValue = curValue;
+                valueChangedHandler(this, args);
+            }
+        }
+
+        private bool OnTouchEventForBgTrack(object source, TouchEventArgs e)
+        {
+            PointStateType state = e.Touch.GetState(0);
+            if (state == PointStateType.Down)
+            {
+                Vector2 pos = e.Touch.GetLocalPosition(0);
+                CalculateCurrentValueByTouch(pos);
+                UpdateValue();
+                if (null != slidingFinishedHandler)
+                {
+                    SlidingFinishedArgs args = new SlidingFinishedArgs();
+                    args.CurrentValue = curValue;
+                    slidingFinishedHandler(this, args);
+                }
+            }
+            return false;
+        }
+
+        private bool OnTouchEventForThumb(object source, TouchEventArgs e)
+        {
+            PointStateType state = e.Touch.GetState(0);
+            if (state == PointStateType.Down)
+            {
+                UpdateState(isFocused, true);
+            }
+            else if (state == PointStateType.Up)
+            {
+                UpdateState(isFocused, false);
+            }
+            return true;
+        }
+
+        private void CalculateCurrentValueByTouch(Vector2 pos)
+        {
+            int bgTrackLength = BgTrackLength();
+            if (direction == DirectionType.Horizontal)
+            {
+                currentSlidedOffset = pos.X;
+            }
+            else if (direction == DirectionType.Vertical)
+            {
+                currentSlidedOffset = bgTrackLength - pos.Y;
+            }
+            if (bgTrackLength != 0)
+            {
+                curValue = ((currentSlidedOffset / (float)bgTrackLength) * (maxValue - minValue)) + minValue;
+                if (null != valueChangedHandler)
+                {
+                    ValueChangedArgs args = new ValueChangedArgs();
+                    args.CurrentValue = curValue;
+                    valueChangedHandler(this, args);
+                }
+            }
+        }
+
+        private void UpdateState(bool isFocusedNew, bool isPressedNew)
+        {
+            if (isFocused == isFocusedNew && isPressed == isPressedNew)
+            {
+                return;
+            }
+            if (thumbImage == null || Style == null)
+            {
+                return;
+            }
+            isFocused = isFocusedNew;
+            isPressed = isPressedNew;
+
+            if (!isFocused && !isPressed)
+            {
+                ControlState = ControlStates.Normal;
+                if (stateChangedHandler != null)
+                {
+                    StateChangedArgs args = new StateChangedArgs();
+                    args.CurrentState = (ControlStates)ControlStates.Normal;
+                    stateChangedHandler(this, args);
+                }
+            }
+            else if (isPressed)
+            {
+                ControlState = ControlStates.Pressed;
+
+                if (stateChangedHandler != null)
+                {
+                    StateChangedArgs args = new StateChangedArgs();
+                    args.CurrentState = (ControlStates)ControlStates.Pressed;
+                    stateChangedHandler(this, args);
+                }
+            }
+            else if (!isPressed && isFocused)
+            {
+                ControlState = ControlStates.Focused;
+
+                if (stateChangedHandler != null)
+                {
+                    StateChangedArgs args = new StateChangedArgs();
+                    args.CurrentState = (ControlStates)ControlStates.Focused;
+                    stateChangedHandler(this, args);
+                }
+            }
+        }
+
+        private void UpdateComponentByIndicatorTypeChanged()
+        {
+            IndicatorType type = CurrentIndicatorType();
+            if (type == IndicatorType.None)
+            {
+                if (lowIndicatorImage != null)
+                {
+                    lowIndicatorImage.Hide();
+                }
+                if (highIndicatorImage != null)
+                {
+                    highIndicatorImage.Hide();
+                }
+                if (lowIndicatorText != null)
+                {
+                    lowIndicatorText.Hide();
+                }
+                if (highIndicatorText != null)
+                {
+                    highIndicatorText.Hide();
+                }
+            }
+            else if (type == IndicatorType.Image)
+            {
+                if (lowIndicatorImage != null)
+                {
+                    lowIndicatorImage.Show();
+                }
+                if (highIndicatorImage != null)
+                {
+                    highIndicatorImage.Show();
+                }
+                if (lowIndicatorText != null)
+                {
+                    lowIndicatorText.Hide();
+                }
+                if (highIndicatorText != null)
+                {
+                    highIndicatorText.Hide();
+                }
+            }
+            else if (type == IndicatorType.Text)
+            {
+                if (lowIndicatorText != null)
+                {
+                    lowIndicatorText.Show();
+                }
+                if (highIndicatorText != null)
+                {
+                    highIndicatorText.Show();
+                }
+                if (lowIndicatorImage != null)
+                {
+                    lowIndicatorImage.Hide();
+                }
+                if (highIndicatorImage != null)
+                {
+                    highIndicatorImage.Hide();
+                }
+            }
+        }
+
+        /// <summary>
+        /// Value Changed event data.
+        /// </summary>
+        /// <since_tizen> 6 </since_tizen>
+        public class ValueChangedArgs : EventArgs
+        {
+            /// <summary>
+            /// Curren value
+            /// </summary>
+            /// <since_tizen> 6 </since_tizen>
+            public float CurrentValue;
+        }
+
+        /// <summary>
+        /// Value Changed event data.
+        /// </summary>
+        /// <since_tizen> 6 </since_tizen>
+        public class SlidingFinishedArgs : EventArgs
+        {
+            /// <summary>
+            /// Curren value
+            /// </summary>
+            /// <since_tizen> 6 </since_tizen>
+            public float CurrentValue;
+        }
+
+        /// <summary>
+        /// State Changed event data.
+        /// </summary>
+        /// <since_tizen> 6 </since_tizen>
+        public class StateChangedArgs : EventArgs
+        {
+            /// <summary>
+            /// Curent state
+            /// </summary>
+            /// <since_tizen> 6 </since_tizen>
+            public ControlStates CurrentState;
+        }
+    }
+}
